Move isEmail helper
