+++ client/views/admin/users/UserForm.js
@@ -12,7 +12,7 @@ import {
 } from '@rocket.chat/fuselage';
 import React, { useCallback, useMemo, useState } from 'react';
 
-import { isEmail } from '../../../../app/utils/lib/isEmail.js';
+import { isEmail } from '../../../../lib/utils/isEmail';
 import CustomFieldsForm from '../../../components/CustomFieldsForm';
 import VerticalBar from '../../../components/VerticalBar';
 import { useTranslation } from '../../../contexts/TranslationContext';
